% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrShiny.R
\name{corrShiny}
\alias{corrShiny}
\title{corrShiny}
\usage{
corrShiny(
  data,
  x_var,
  y_var,
  color_var = NULL,
  size_var = NULL,
  correlation_method = "pearson"
)
}
\arguments{
\item{data}{A data frame with the variables to be analyzed.}

\item{x_var}{The name of the variable to be plotted on the X-axis.}

\item{y_var}{The name of the variable to be plotted on the Y-axis.}

\item{color_var}{The name of the variable to be used for coloring the points on the scatter plot.}

\item{size_var}{The name of the variable to be used for sizing the points on the scatter plot.}

\item{correlation_method}{The method to be used for computing the correlation coefficient, must be one of "pearson", "spearman" or "kendall".}
}
\value{
A Shiny app that displays a scatter plot and the correlation coefficient between two variables.
}
\description{
Correlation Explorer Shiny App. This function creates
a Shiny application to explore the correlation between
variables in a given dataset.
}
