% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animSolar.R
\name{animSolar}
\alias{animSolar}
\title{animSolar}
\usage{
animSolar(
  data,
  method = c("pearson", "kendall", "spearman"),
  sun = NULL,
  path = NULL,
  gif_name = "solar_system.gif",
  fps = 60
)
}
\arguments{
\item{data}{A data frame.}

\item{method}{Which correlation coefficient (or covariance) is to be computed.
One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{sun}{A variable name to be chosen as the dependent variable.}

\item{path}{Path of the directory to save plot to. Defaults to the working directory.}

\item{gif_name}{File name to create on disk. should be in the format "myfile.gif"}

\item{fps}{The frame rate of the animation in frames/sec. default is 60 fps.}
}
\value{
An animated solar system plot displaying correlations.
}
\description{
Creates an animated solar system correlation plot.
}
\details{
In a solar system correlation plot, the dependent variable of
interest is positioned at the center, represented as the sun.
The explanatory variables are depicted as planets orbiting
around the sun, with their distance from the sun corresponding
to the absolute value of their correlation with the dependent variable.
Therefore, the greater the distance of a planet from the sun,
the weaker the correlation between the explanatory variable
and the dependent variable.
}
