% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrBarplot.R
\name{corrBarplot}
\alias{corrBarplot}
\title{corrBarplot}
\usage{
corrBarplot(
  data,
  method = c("pearson", "kendall", "spearman"),
  interactive = FALSE,
  pal = colorRampPalette(c("cornflowerblue", "white", "tomato"))(100)
)
}
\arguments{
\item{data}{A data frame.}

\item{method}{Which correlation coefficient (or covariance) is to be computed.
One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{interactive}{If TRUE then an interacive version of the barplot is displayed.}

\item{pal}{The colour palette to use for displaying values.}
}
\value{
A static or interactive bar plot displaying correlations.
}
\description{
Creates either a static or interactive bar plot displaying correlations.
}
\details{
Creates a static or interactive bar plot displaying correlation values. By hovering
mouse over a bar, the variables and correlation value is shown.
}
\examples{
corrHeatmap(data = mtcars, display = 'all')


}
