% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrBubble.R
\name{corrBubble}
\alias{corrBubble}
\title{corrBubble}
\usage{
corrBubble(
  data,
  method = c("pearson", "kendall", "spearman"),
  display = c("all", "upper", "lower"),
  pal = colorRampPalette(c("cornflowerblue", "white", "tomato"))(100)
)
}
\arguments{
\item{data}{A data frame.}

\item{method}{Which correlation coefficient (or covariance) is to be computed.
One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{display}{Choose to display the upper, lower, or full martix of vaules using
"upper", "lower", or "all", respectively.}

\item{pal}{The colour palette to use for displaying values.}
}
\value{
An interactive bubble plot displaying correlations.
}
\description{
Creates an interactive bubble plot displaying correlations.
}
\details{
Creates an interactive bubble plot displaying correlation values. By hovering
mouse over a cell, the variables and correlation value is shown.
}
\examples{
corrBubble(data = mtcars, display = 'all')


}
