% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrChord.R
\name{corrChord}
\alias{corrChord}
\title{corrChord}
\usage{
corrChord(
  data,
  method = c("pearson", "kendall", "spearman"),
  threshold = 0,
  circle = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the data to be analyzed.}

\item{method}{A character string specifying the correlation method. One of
"pearson", "kendall", or "spearman". Default is "pearson".}

\item{threshold}{A numeric value indicating the minimum absolute
correlation value to display in the plot.}

\item{circle}{A logical value indicating whether to use a circular
layout (TRUE) or linear layout (FALSE), default is FALSE.}
}
\value{
A chord plot displaying correlations.
}
\description{
This function creates a chord plot of correlations between variables in a dataset.
}
\details{
When using a large amount of data, this plot can quickly become over
complicated. It is recommended to filter the correlations using the \code{threshold}
argument to simplify the visualisation.
}
\examples{
corrChord(data = mtcars,
          method = 'pearson',
          threshold = 0.8)

corrChord(data = mtcars,
          method = 'pearson',
          threshold = 0.8,
          circle = TRUE)

}
