% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrChord.R
\name{corrChord}
\alias{corrChord}
\title{corrChord}
\usage{
corrChord(
  data,
  method = c("pearson", "kendall", "spearman"),
  threshold = 0,
  circle = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{method}{Which correlation coefficient (or covariance) is to be computed.
One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{threshold}{Filter correlations with an absolute value lower than selected value.}

\item{circle}{If TRUE then plot is displayed as a circle.}
}
\value{
A chord plot displaying correlations.
}
\description{
Creates an chord plot displaying correlations.
}
\details{
DEtS hEre!!!
}
\examples{
corrChord(data = mtcars,
          method = 'pearson',
          threshold = 0.8)

corrChord(data = mtcars,
          method = 'pearson',
          threshold = 0.8,
          circle = TRUE)

}
