% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrCircle.R
\name{corrCircle}
\alias{corrCircle}
\title{corrCircle}
\usage{
corrCircle(
  data,
  method = c("pearson", "kendall", "spearman"),
  threshold = 0,
  ticks = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{method}{Which correlation coefficient (or covariance) is to be computed.
One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{threshold}{Filter correlations with an absolute value greater than selected value.
#'}
}
\value{
A circular chord plot displaying correlations.
}
\description{
Creates a circular chord plot displaying correlations.
}
\details{
DEtS hEre!!!
}
