% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrNetwork.R
\name{corrNetwork}
\alias{corrNetwork}
\title{corrNetwork}
\usage{
corrNetwork(
  data,
  method = c("pearson", "kendall", "spearman"),
  threshold = 0,
  layout = "layout_nicely",
  width = "100\%",
  height = "400px",
  physics = TRUE
)
}
\arguments{
\item{data}{A data frame or matrix containing the input data for the correlation calculation.}

\item{method}{A character string specifying the correlation method. One of
"pearson", "kendall", or "spearman". Default is "pearson".}

\item{threshold}{A numeric value indicating the minimum absolute correlation
value to display in the plot.}

\item{layout}{Any \code{igraph} layout to display the network.}

\item{width}{The width of the viewing window.}

\item{height}{The height of the viewing window.}

\item{physics}{A logical value indicating whether to use physics-based layout. Default is TRUE.}
}
\value{
A network plot displaying correlations.
}
\description{
Creates an interactive Correlation Network Visualization
}
\details{
Each node in the network represents a variable where the width of
the connecting edges represent the absolute value of the correlation. Positive
correlations have red coloured edges whereas negative correlations have blue coloured
edges.
}
\examples{
corrNetwork(data = iris[,1:4], threshold = 0.5)

corrNetwork(data = mtcars,
           threshold = 0.8,
           method = "pearson",
           layout = 'layout_on_grid',
           physics = FALSE)
}
