% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrNetwork.R
\name{corrNetwork}
\alias{corrNetwork}
\title{corrNetwork}
\usage{
corrNetwork(
  data,
  method = c("pearson", "kendall", "spearman"),
  filterCorr = 0,
  layout = "layout_nicely",
  width = "100\%",
  height = "400px",
  physics = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{method}{Which correlation coefficient (or covariance) is to be computed.
One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{filterCorr}{Filter correlations with an absolute value greater than selected value.}

\item{layout}{Use an \code{igraph} layout to display network.}

\item{width}{The width of the viewing window.}

\item{height}{The height of the viewing window.}

\item{physics}{If TRUE (the default) then physics is enabled on nodes. This will
alter the layout to be roughly circular.}
}
\value{
A network plot displaying correlations.
}
\description{
Creates an interactive network plot displaying correlations.
}
\details{
Each node in the network represents a variable where the width of
the connecting edges represent the absolute value of the correlation. Positive
correlations have red coloured edges whereas negative correlations have blue coloured
edges.
}
\examples{
corrNetwork(data = iris[,1:4], filterCorr = 0.5)

library(MASS)
data("Boston")
corrNetwork(data = Boston,
           filterCorr = 0.5,
            method = "kendall",
           layout = 'layout_on_grid',
           physics = F)
}
