% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrSolar.R
\name{corrSolar}
\alias{corrSolar}
\title{corrSolar}
\usage{
corrSolar(data, method = c("pearson", "kendall", "spearman"), sun = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{method}{Which correlation coefficient (or covariance) is to be computed.
One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{sun}{A variable name to be chosen as the dependent variable.}
}
\value{
An solar system plot displaying correlations.
}
\description{
Creates a solar system correlation plot.
}
\details{
In a solar system correlation plot, the dependent variable of
interest is positioned at the center, represented as the sun.
The explanatory variables are depicted as planets orbiting
around the sun, with their distance from the sun corresponding
to the absolute value of their correlation with the dependent variable.
Therefore, the greater the distance of a planet from the sun,
the weaker the correlation between the explanatory variable
and the dependent variable.
}
\examples{
corrSolar(mtcars, sun = 'mpg')


}
